requirements_extraction.txt
pymupdf
pdfplumber
pillow
